+++ pgdrive/envs/generation_envs/change_density_env.py
@@ -1,7 +1,7 @@
 import logging
 
 from pgdrive.envs.pgdrive_env import PGDriveEnv
-from pgdrive.pg_config import PGConfig
+from pgdrive import PGConfig
 from pgdrive.utils import get_np_random
