+++ pgdrive/envs/generation_envs/safe_pgdrive_env.py
@@ -1,5 +1,5 @@
 from pgdrive.envs.pgdrive_env import PGDriveEnv
-from pgdrive.pg_config import PGConfig
+from pgdrive import PGConfig
 from pgdrive.rl_utils.done import pg_done_function
