+++ pgdrive/envs/pgdrive_env_v2.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from pgdrive.constants import DEFAULT_AGENT
 from pgdrive.envs.pgdrive_env import PGDriveEnv as PGDriveEnvV1
-from pgdrive.pg_config import PGConfig
+from pgdrive import PGConfig
 from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 from pgdrive.scene_manager.traffic_manager import TrafficMode
 from pgdrive.utils import clip, merge_dicts
