+++ pgdrive/scene_creator/blocks/straight.py
@@ -1,5 +1,4 @@
-from pgdrive.pg_config.parameter_space import BlockParameterSpace, Parameter
-from pgdrive.utils.pg_space import PGSpace
+from pgdrive.utils.pg_space import PGSpace, Parameter, BlockParameterSpace
 from pgdrive.scene_creator.blocks.block import Block, BlockSocket
 from pgdrive.scene_creator.blocks.create_block_utils import ExtendStraightLane, CreateRoadFrom, CreateAdverseRoad
 from pgdrive.scene_creator.lane.abs_lane import LineType
