+++ pgdrive/scene_creator/map.py
@@ -7,8 +7,8 @@ from typing import List
 import numpy as np
 from panda3d.core import NodePath
 
-from pgdrive.pg_config import PGConfig
-from pgdrive.pg_config.pg_blocks import PGBlock
+from pgdrive import PGConfig
+from pgdrive.scene_creator.pg_blocks import PGBlock
 from pgdrive.scene_creator.algorithm.BIG import BIG, BigGenerateMethod
 from pgdrive.scene_creator.blocks.block import Block
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
