+++ pgdrive/scene_creator/vehicle_module/routing_localization.py
@@ -3,7 +3,7 @@ import logging
 import numpy as np
 from panda3d.core import BitMask32, LQuaternionf, TransparencyAttrib
 from pgdrive.constants import COLLISION_INFO_COLOR, RENDER_MODE_ONSCREEN, CamMask
-from pgdrive.pg_config.parameter_space import BlockParameterSpace, Parameter
+from pgdrive.utils.pg_space import Parameter, BlockParameterSpace
 from pgdrive.scene_creator.blocks.first_block import FirstBlock
 from pgdrive.scene_creator.lane.circular_lane import CircularLane
 from pgdrive.scene_creator.map import Map
