+++ pgdrive/utils/__init__.py
@@ -1,6 +1,7 @@
 from pgdrive.utils.asset_loader import AssetLoader, initialize_asset_loader
 from pgdrive.utils.draw_top_down_map import draw_top_down_map
 from pgdrive.utils.math_utils import safe_clip, clip, norm
+from pgdrive.utils.pg_config import PGConfig
 from pgdrive.utils.random import get_np_random, RandomEngine
 from pgdrive.utils.utils import is_mac, import_pygame, recursive_equal, setup_logger, random_string, merge_dicts, \
     concat_step_infos
