+++ pgdrive/utils/pg_space.py
@@ -1,11 +1,152 @@
+import logging
 import typing as tp
-from collections import namedtuple
+from collections import namedtuple, OrderedDict
 
-from pgdrive.pg_config.space import Dict, Box, Discrete
+import numpy as np
+from pgdrive.utils import get_np_random
 
 PGBoxSpace = namedtuple("PGBoxSpace", "max min")
 PGDiscreteSpace = namedtuple("PGDiscreteSpace", "number")
 PGConstantSpace = namedtuple("PGConstantSpace", "value")
+"""
+This filed is mostly copied from gym==0.17.2
+We use the gym.spaces as helpers, but it may cause problem if user using some old version of gym.
+"""
+
+
+class Space:
+    """
+    Copied from gym: gym/spaces/space.py
+
+    Defines the observation and action spaces, so you can write generic
+    code that applies to any Env. For example, you can choose a random
+    action.
+    """
+
+    def __init__(self, shape=None, dtype=None):
+        import numpy as np  # takes about 300-400ms to import, so we load lazily
+        self.shape = None if shape is None else tuple(shape)
+        self.dtype = None if dtype is None else np.dtype(dtype)
+        self.np_random = None
+        self.seed()
+
+    def sample(self):
+        """Randomly sample an element of this space. Can be
+        uniform or non-uniform sampling based on boundedness of space."""
+        raise NotImplementedError
+
+    def seed(self, seed=None):
+        """Seed the PRNG of this space. """
+        self.np_random, seed = get_np_random(seed, return_seed=True)
+        return [seed]
+
+    def contains(self, x):
+        """
+        Return boolean specifying if x is a valid
+        member of this space
+        """
+        raise NotImplementedError
+
+    def __contains__(self, x):
+        return self.contains(x)
+
+    def to_jsonable(self, sample_n):
+        """Convert a batch of samples from this space to a JSONable data type."""
+        # By default, assume identity is JSONable
+        return sample_n
+
+    def from_jsonable(self, sample_n):
+        """Convert a JSONable data type to a batch of samples from this space."""
+        # By default, assume identity is JSONable
+        return sample_n
+
+
+class Dict(Space):
+    """
+    Copied from gym: gym/spaces/dcit.py
+
+    A dictionary of simpler spaces.
+
+    Example usage:
+    self.observation_space = spaces.Dict({"position": spaces.Discrete(2), "velocity": spaces.Discrete(3)})
+
+    Example usage [nested]:
+    self.nested_observation_space = spaces.Dict({
+        'sensors':  spaces.Dict({
+            'position': spaces.Box(low=-100, high=100, shape=(3,)),
+            'velocity': spaces.Box(low=-1, high=1, shape=(3,)),
+            'front_cam': spaces.Tuple((
+                spaces.Box(low=0, high=1, shape=(10, 10, 3)),
+                spaces.Box(low=0, high=1, shape=(10, 10, 3))
+            )),
+            'rear_cam': spaces.Box(low=0, high=1, shape=(10, 10, 3)),
+        }),
+        'ext_controller': spaces.MultiDiscrete((5, 2, 2)),
+        'inner_state':spaces.Dict({
+            'charge': spaces.Discrete(100),
+            'system_checks': spaces.MultiBinary(10),
+            'job_status': spaces.Dict({
+                'task': spaces.Discrete(5),
+                'progress': spaces.Box(low=0, high=100, shape=()),
+            })
+        })
+    })
+    """
+
+    def __init__(self, spaces=None, **spaces_kwargs):
+        assert (spaces is None) or (not spaces_kwargs), 'Use either Dict(spaces=dict(...)) or Dict(foo=x, bar=z)'
+        if spaces is None:
+            spaces = spaces_kwargs
+        if isinstance(spaces, dict) and not isinstance(spaces, OrderedDict):
+            spaces = OrderedDict(sorted(list(spaces.items())))
+        if isinstance(spaces, list):
+            spaces = OrderedDict(spaces)
+        self.spaces = spaces
+        for space in spaces.values():
+            assert isinstance(space, Space), 'Values of the dict should be instances of gym.Space'
+        super(Dict, self).__init__(None, None)  # None for shape and dtype, since it'll require special handling
+
+    def seed(self, seed=None):
+        [space.seed(seed) for space in self.spaces.values()]
+
+    def sample(self):
+        return OrderedDict([(k, space.sample()) for k, space in self.spaces.items()])
+
+    def contains(self, x):
+        if not isinstance(x, dict) or len(x) != len(self.spaces):
+            return False
+        for k, space in self.spaces.items():
+            if k not in x:
+                return False
+            if not space.contains(x[k]):
+                return False
+        return True
+
+    def __getitem__(self, key):
+        return self.spaces[key]
+
+    def __repr__(self):
+        return "Dict(" + ", ".join([str(k) + ":" + str(s) for k, s in self.spaces.items()]) + ")"
+
+    def to_jsonable(self, sample_n):
+        # serialize as dict-repr of vectors
+        return {key: space.to_jsonable([sample[key] for sample in sample_n]) \
+                for key, space in self.spaces.items()}
+
+    def from_jsonable(self, sample_n):
+        dict_of_list = {}
+        for key, space in self.spaces.items():
+            dict_of_list[key] = space.from_jsonable(sample_n[key])
+        ret = []
+        for i, _ in enumerate(dict_of_list[key]):
+            entry = {}
+            for key, value in dict_of_list.items():
+                entry[key] = value[i]
+            ret.append(entry)
+        return ret
+
+    def __eq__(self, other):
+        return isinstance(other, Dict) and self.spaces == other.spaces
 
 
 class PGSpace(Dict):
@@ -14,6 +155,7 @@ class PGSpace(Dict):
     Usage:
     PGSpace({"lane_length":length})
     """
+
     def __init__(self, our_config: tp.Dict[str, tp.Union[PGBoxSpace, PGDiscreteSpace, PGConstantSpace]]):
         super(PGSpace, self).__init__(PGSpace.wrap2gym_space(our_config))
         self.parameters = set(our_config.keys())
@@ -23,14 +165,297 @@ class PGSpace(Dict):
         ret = dict()
         for key, value in our_config.items():
             if isinstance(value, PGBoxSpace):
-                ret[key] = Box(low=value.min, high=value.max, shape=(1, ))
+                ret[key] = Box(low=value.min, high=value.max, shape=(1,))
             elif isinstance(value, PGDiscreteSpace):
                 ret[key] = Discrete(value.number)
             elif isinstance(value, PGConstantSpace):
-                ret[key] = Box(low=value.value, high=value.value, shape=(1, ))
+                ret[key] = Box(low=value.value, high=value.value, shape=(1,))
             else:
                 raise ValueError("{} can not be wrapped in gym space".format(key))
         return ret
+
+
+class Parameter:
+    """
+    Block parameters and vehicle parameters
+    """
+    # block
+    length = "length"
+    radius = "radius"
+    angle = "angle"
+    goal = "goal"
+    dir = "dir"
+    radius_inner = "inner_radius"  # only for roundabout use
+    radius_exit = "exit_radius"
+    t_intersection_type = "t_type"
+    lane_num = "lane_num"
+    change_lane_num = "change_lane_num"
+    decrease_increase = "decrease_increase"
+
+    # vehicle
+    vehicle_length = "v_len"
+    vehicle_width = "v_width"
+    vehicle_height = "v_height"
+    front_tire_longitude = "f_tire_long"
+    rear_tire_longitude = "r_tire_long"
+    tire_lateral = "tire_lateral"
+    tire_axis_height = "tire_axis_height"
+    tire_radius = "tire_radius"
+    mass = "mass"
+    chassis_height = "chassis_height"
+    heading = "heading"
+    steering_max = "steering_max"
+    engine_force_max = "e_f_max"
+    brake_force_max = "b_f_max"
+    speed_max = "s_max"
+
+    # vehicle visualization
+    vehicle_vis_z = "vis_z"
+    vehicle_vis_y = "vis_y"
+    vehicle_vis_h = "vis_h"
+    vehicle_vis_scale = "vis_scale"
+
+
+class VehicleParameterSpace:
+    BASE_VEHICLE = {
+        # Now the parameter sample is not available and thus the value space is incorrect
+        Parameter.vehicle_length: PGConstantSpace(4.0),
+        Parameter.vehicle_width: PGConstantSpace(1.5),
+        Parameter.vehicle_height: PGConstantSpace(1),
+        Parameter.chassis_height: PGConstantSpace(0.3),
+        Parameter.front_tire_longitude: PGConstantSpace(1.05),
+        Parameter.rear_tire_longitude: PGConstantSpace(1.17),
+        Parameter.tire_lateral: PGConstantSpace(0.8),
+        Parameter.tire_radius: PGConstantSpace(0.25),
+        Parameter.mass: PGConstantSpace(800.0),
+        Parameter.heading: PGConstantSpace(0.0),
+
+        # visualization
+        Parameter.vehicle_vis_h: PGConstantSpace(180),
+        Parameter.vehicle_vis_y: PGConstantSpace(0.1),
+        Parameter.vehicle_vis_z: PGConstantSpace(-0.31),
+        Parameter.vehicle_vis_scale: PGConstantSpace(0.013),
+
+        # TODO the following parameters will be opened soon using PGBoxSPace
+        Parameter.steering_max: PGConstantSpace(40.0),
+        Parameter.engine_force_max: PGConstantSpace(500.0),
+        Parameter.brake_force_max: PGConstantSpace(40.0),
+        Parameter.speed_max: PGConstantSpace(120),
+    }
+
+
+class BlockParameterSpace:
+    """
+    Make sure the range of curve parameters covers the parameter space of other blocks,
+    otherwise, an error may happen in navigation info normalization
+    """
+    STRAIGHT = {Parameter.length: PGBoxSpace(min=40.0, max=80.0)}
+    CURVE = {
+        Parameter.length: PGBoxSpace(min=40.0, max=80.0),
+        Parameter.radius: PGBoxSpace(min=25.0, max=60.0),
+        Parameter.angle: PGBoxSpace(min=45, max=135),
+        Parameter.dir: PGDiscreteSpace(2)
+    }
+    INTERSECTION = {
+        Parameter.radius: PGConstantSpace(10),
+        Parameter.change_lane_num: PGDiscreteSpace(number=2),  # 0, 1
+        Parameter.decrease_increase: PGDiscreteSpace(number=2)  # 0, decrease, 1 increase
+    }
+    ROUNDABOUT = {
+        Parameter.radius_exit: PGBoxSpace(min=5, max=15),  # TODO Should we reduce this?
+        Parameter.radius_inner: PGBoxSpace(min=15, max=45),  # TODO Should we reduce this?
+        Parameter.angle: PGConstantSpace(60)
+    }
+    T_INTERSECTION = {
+        Parameter.radius: PGConstantSpace(10),
+        Parameter.t_intersection_type: PGDiscreteSpace(number=3),  # 3 different t type for previous socket
+        Parameter.change_lane_num: PGDiscreteSpace(2),  # 0,1
+        Parameter.decrease_increase: PGDiscreteSpace(2)  # 0, decrease, 1 increase
+    }
+    RAMP_PARAMETER = {
+        Parameter.length: PGBoxSpace(min=20, max=40)  # accelerate/decelerate part length
+    }
+    FORK_PARAMETER = {
+        Parameter.length: PGBoxSpace(min=20, max=40),  # accelerate/decelerate part length
+        Parameter.lane_num: PGDiscreteSpace(2)
+    }
+    BOTTLENECK_PARAMETER = {
+        Parameter.length: PGBoxSpace(min=20, max=50),  # the length of straigh part
+        Parameter.lane_num: PGDiscreteSpace(2),  # the lane num increased or descreased now fix to 1
+    }
+
+
+class Discrete(Space):
+    r"""
+    Copied from gym: gym/spaces/discrete.py
+
+    A discrete space in :math:`\{ 0, 1, \\dots, n-1 \}`.
+
+    Example::
+
+        >>> Discrete(2)
+
+    """
+
+    def __init__(self, n):
+        assert n >= 0
+        self.n = n
+        super(Discrete, self).__init__((), np.int64)
+
+    def sample(self):
+        return self.np_random.randint(self.n)
+
+    def contains(self, x):
+        if isinstance(x, int):
+            as_int = x
+        elif isinstance(x, (np.generic, np.ndarray)) and (x.dtype.char in np.typecodes['AllInteger'] and x.shape == ()):
+            as_int = int(x)
+        else:
+            return False
+        return as_int >= 0 and as_int < self.n
+
+    def __repr__(self):
+        return "Discrete(%d)" % self.n
+
+    def __eq__(self, other):
+        return isinstance(other, Discrete) and self.n == other.n
+
+
+class Box(Space):
+    """
+    Copied from gym: gym/spaces/box.py
+
+    A (possibly unbounded) box in R^n. Specifically, a Box represents the
+    Cartesian product of n closed intervals. Each interval has the form of one
+    of [a, b], (-oo, b], [a, oo), or (-oo, oo).
+
+    There are two common use cases:
+
+    * Identical bound for each dimension::
+        >>> Box(low=-1.0, high=2.0, shape=(3, 4), dtype=np.float32)
+        Box(3, 4)
+
+    * Independent bound for each dimension::
+        >>> Box(low=np.array([-1.0, -2.0]), high=np.array([2.0, 4.0]), dtype=np.float32)
+        Box(2,)
+
+    """
+
+    def __init__(self, low, high, shape=None, dtype=np.float32):
+        assert dtype is not None, 'dtype must be explicitly provided. '
+        self.dtype = np.dtype(dtype)
+
+        # determine shape if it isn't provided directly
+        if shape is not None:
+            shape = tuple(shape)
+            assert np.isscalar(low) or low.shape == shape, "low.shape doesn't match provided shape"
+            assert np.isscalar(high) or high.shape == shape, "high.shape doesn't match provided shape"
+        elif not np.isscalar(low):
+            shape = low.shape
+            assert np.isscalar(high) or high.shape == shape, "high.shape doesn't match low.shape"
+        elif not np.isscalar(high):
+            shape = high.shape
+            assert np.isscalar(low) or low.shape == shape, "low.shape doesn't match high.shape"
+        else:
+            raise ValueError("shape must be provided or inferred from the shapes of low or high")
+
+        if np.isscalar(low):
+            low = np.full(shape, low, dtype=dtype)
+
+        if np.isscalar(high):
+            high = np.full(shape, high, dtype=dtype)
+
+        self.shape = shape
+        self.low = low
+        self.high = high
+
+        def _get_precision(dtype):
+            if np.issubdtype(dtype, np.floating):
+                return np.finfo(dtype).precision
+            else:
+                return np.inf
+
+        low_precision = _get_precision(self.low.dtype)
+        high_precision = _get_precision(self.high.dtype)
+        dtype_precision = _get_precision(self.dtype)
+        if min(low_precision, high_precision) > dtype_precision:
+            logging.warning("Box bound precision lowered by casting to {}".format(self.dtype))
+        self.low = self.low.astype(self.dtype)
+        self.high = self.high.astype(self.dtype)
+
+        # Boolean arrays which indicate the interval type for each coordinate
+        self.bounded_below = -np.inf < self.low
+        self.bounded_above = np.inf > self.high
+
+        super(Box, self).__init__(self.shape, self.dtype)
+
+    def is_bounded(self, manner="both"):
+        below = np.all(self.bounded_below)
+        above = np.all(self.bounded_above)
+        if manner == "both":
+            return below and above
+        elif manner == "below":
+            return below
+        elif manner == "above":
+            return above
+        else:
+            raise ValueError("manner is not in {'below', 'above', 'both'}")
+
+    def sample(self):
+        """
+        Generates a single random sample inside of the Box.
+
+        In creating a sample of the box, each coordinate is sampled according to
+        the form of the interval:
+
+        * [a, b] : uniform distribution
+        * [a, oo) : shifted exponential distribution
+        * (-oo, b] : shifted negative exponential distribution
+        * (-oo, oo) : normal distribution
+        """
+        high = self.high if self.dtype.kind == 'f' \
+            else self.high.astype('int64') + 1
+        sample = np.empty(self.shape)
+
+        # Masking arrays which classify the coordinates according to interval
+        # type
+        unbounded = ~self.bounded_below & ~self.bounded_above
+        upp_bounded = ~self.bounded_below & self.bounded_above
+        low_bounded = self.bounded_below & ~self.bounded_above
+        bounded = self.bounded_below & self.bounded_above
+
+        # Vectorized sampling by interval type
+        sample[unbounded] = self.np_random.normal(size=unbounded[unbounded].shape)
+
+        sample[low_bounded] = self.np_random.exponential(size=low_bounded[low_bounded].shape) + self.low[low_bounded]
+
+        sample[upp_bounded] = -self.np_random.exponential(size=upp_bounded[upp_bounded].shape) + self.high[upp_bounded]
+
+        sample[bounded] = self.np_random.uniform(low=self.low[bounded], high=high[bounded], size=bounded[bounded].shape)
+        if self.dtype.kind == 'i':
+            sample = np.floor(sample)
+
+        return sample.astype(self.dtype)
+
+    def contains(self, x):
+        if isinstance(x, list):
+            x = np.array(x)  # Promote list to array for contains check
+        return x.shape == self.shape and np.all(x >= self.low) and np.all(x <= self.high)
+
+    def to_jsonable(self, sample_n):
+        return np.array(sample_n).tolist()
+
+    def from_jsonable(self, sample_n):
+        return [np.asarray(sample) for sample in sample_n]
+
+    def __repr__(self):
+        return "Box" + str(self.shape)
+
+    def __eq__(self, other):
+        return isinstance(other, Box) and \
+               (self.shape == other.shape) and \
+               np.allclose(self.low, other.low) and \
+               np.allclose(self.high, other.high)
 
 
 if __name__ == "__main__":
